refactor: move ClientMigrationHandler to correct package
