+++ migration/identity-migration/src/main/java/io/camunda/migration/identity/config/MigrationHandlerConfig.java
@@ -8,12 +8,12 @@
 package io.camunda.migration.identity.config;
 
 import io.camunda.migration.identity.AuthorizationMigrationHandler;
+import io.camunda.migration.identity.ClientMigrationHandler;
 import io.camunda.migration.identity.GroupMigrationHandler;
 import io.camunda.migration.identity.StaticConsoleRoleAuthorizationMigrationHandler;
 import io.camunda.migration.identity.StaticConsoleRoleMigrationHandler;
 import io.camunda.migration.identity.console.ConsoleClient;
 import io.camunda.migration.identity.midentity.ManagementIdentityClient;
-import io.camunda.migration.identity.transformer.ClientMigrationHandler;
 import io.camunda.security.auth.Authentication;
 import io.camunda.service.AuthorizationServices;
 import io.camunda.service.GroupServices;
